Chef is fond of burgers
Problem
Chef is fond of burgers and decided to make as many burgers as possible.

Chef has
�
A patties and
�
B buns. To make
1
1 burger, Chef needs
1
1 patty and
1
1 bun.
Find the maximum number of burgers that Chef can make.

Input Format
The first line of input will contain an integer
�
T — the number of test cases. The description of
�
T test cases follows.
The first and only line of each test case contains two space-separated integers
�
A and
�
B, the number of patties and buns respectively.
Output Format
For each test case, output the maximum number of burgers that Chef can make.

Constraints
1
≤
�
≤
1000
1≤T≤1000
1
≤
�
,
�
≤
1
0
5
1≤A,B≤10
5