Audible Range
Problem
Chef's dog binaryhears frequencies starting from 67 Hertz to 45000 Hertz (both inclusive).
If Chefs commands have a frequency of X Hertz, find whether binarycan hear them or not.
Input Format
The first line of input will contain a single integer T, denoting the number of test cases.
• Each test case consists of a single integer X - the frequency of Chefs commands in Hertz.
Output Format
For each test case, output on a new line YES, if binarycan hear Chefs commands. Otherwise, print NO.
The output is case-insensitive. Thus, the strings YES, yes, yes, and Yes are all considered the same.

Explanation:
Test case 1: Chefs command has a frequency of 42 Hertz which is less than 67. Thus, it would not be
audible to binary.
Test case 2: Chefs command has a frequency of 67 Hertz which lies in the range [67, 45000]. Thus, it
would be audible to binary.
Test case 3: Chefs command has a frequency of 402 Hertz which lies in the range [67, 45000]. Thus, it
would be audible to binary.
Test case 4: Chefs command has a frequency of 45000 Hertz which lies in the range [67, 45000]. Thus,
it would be audible to binary.
Test case 5: Chefs command has a frequency of 45005 Hertz which is greater than 45000. Thus, it
would not be audible to binary.